More Things!
The Magazine is now an AmmoStack, so the first bullet you put in will be
the last one out. I can also make it so a magazine can only take the
ammo for the type of weapon it's used for (a handgun magazine can't take
a shotgun shell). Also, I changed the Ammo class to represent a single
round, storing the ammo type and it's status effect.