Turn Order using List.OrderByDescending!
Okay, so in order to get the turn order working properly I decided to
use List.OrderByDescending instead of the algorithm I made (which was
essentially a poor implementation of the method provided by List).  I
noticed that for some reason the Zeds are taking their turn before the
players, which they use to surround the player, and I'm having
difficulty figuring out why, so if any of you could help, that would be
much appreciated.