Rots of Rittle Fixes.
Fixed the ToString methods of the player classes because it was
bothering me. Made sure to set the Name property for ammo. Changed some
clauses that say "x.GetType == blah" to "x is blah". Had to implement
some classes for Trader so VS would stop yelling at me, and lastly I
altered the ToString of character.cs so instead of explicity stating the
class, class is now determined dynamically with this.GetType().Name.