Bug Fix
Fixed a bug where the game would crash due to a nullpointerexception
when the controller cleaned the dead bodies from the field. If the dead
character is a zed, inventory is null, so when the inventory is dumped
at death, it would crash. Just put in a simple check for a null
inventory to fix it.